+++ src/hub/interfaces/IHub.sol
@@ -52,8 +52,8 @@ interface IHub is IHubBase, IAccessManaged {
     uint128 drawnShares;
     //
     uint128 addedShares;
-    uint56 addCap;
-    uint56 drawCap;
+    uint40 addCap;
+    uint40 drawCap;
     bool active;
     bool paused;
     //
@@ -63,8 +63,8 @@ interface IHub is IHubBase, IAccessManaged {
   struct SpokeConfig {
     bool active;
     bool paused;
-    uint56 addCap;
-    uint56 drawCap;
+    uint40 addCap;
+    uint40 drawCap;
   }
 
   /// @notice Emitted when an asset is added.
@@ -377,7 +377,7 @@ interface IHub is IHubBase, IAccessManaged {
   /// @notice Returns the maximum value for any spoke cap (add or draw).
   /// @dev The value is not inclusive; using the maximum value indicates no cap.
   /// @return The maximum cap value, expressed in asset units.
-  function MAX_ALLOWED_SPOKE_CAP() external view returns (uint56);
+  function MAX_ALLOWED_SPOKE_CAP() external view returns (uint40);
 
   /// @notice Returns the maximum allowed risk premium.
   /// @dev This is used to validate and limit premium values reported by spokes.
