+++ tests/unit/Spoke/Spoke.MultipleHub.SiloedBorrowing.t.sol
@@ -8,8 +8,8 @@ contract SpokeMultipleHubSiloedBorrowingTest is SpokeMultipleHubBase {
   struct SiloedLocalVars {
     uint256 assetAId;
     uint256 assetBId;
-    uint56 assetAAddCap;
-    uint56 assetBDrawCap;
+    uint40 assetAAddCap;
+    uint40 assetBDrawCap;
     uint256 reserveAId;
     uint256 reserveBId;
     uint256 reserveAIdNewSpoke;
